some context and security improvement
improved some context and added security to deletion. it will not try to delete all keys if an empty value is given, or accidental key pressed

+++ Context Menu Remover (run as admin).bat
@@ -2,6 +2,8 @@
 echo.[1] custom mode [2] normal mode
 choice /c 12
 echo.Please DONT put nothing as an argument, it'll mean to delete the entire thing.
+echo.
+echo.For "normal" mode, strip "01" from name: "01MenuCommand" is to be "MenuCommand"
 IF %ERRORLEVEL% == 1 goto custom:
 IF %ERRORLEVEL% == 2 goto normal
 echo. ERROR: Invalid choice
@@ -9,37 +11,59 @@ pause
 exit
 
 :normal
-echo.Whish value do you wanna remove?
+echo.---
+echo.Which value do you wanna remove?
 reg query "HKEY_CLASSES_ROOT\Directory\ContextMenus"
+echo...
 echo.
+set value=
 set /p value=
-reg delete "HKEY_CLASSES_ROOT\Directory\ContextMenus\%value%"
-reg delete "HKEY_CLASSES_ROOT\Directory\Background\shell\01%value%"
-reg delete "HKEY_CLASSES_ROOT\Directory\Background\shell\%value%"
-reg delete "HKEY_CLASSES_ROOT\Directory\shell\01%value%"
-reg delete "HKEY_CLASSES_ROOT\Directory\shell\%value%"
+IF "%value%" == "" GOTO empty
+IF NOT "%value%" == "" reg delete "HKEY_CLASSES_ROOT\Directory\ContextMenus\%value%"
+IF NOT "%value%" == "" reg delete "HKEY_CLASSES_ROOT\Directory\Background\shell\01%value%"
+IF NOT "%value%" == "" reg delete "HKEY_CLASSES_ROOT\Directory\Background\shell\%value%"
+IF NOT "%value%" == "" reg delete "HKEY_CLASSES_ROOT\Directory\shell\01%value%"
+IF NOT "%value%" == "" reg delete "HKEY_CLASSES_ROOT\Directory\shell\%value%"
 echo.
 echo.all done!
 pause
 exit
 
 :custom
-echo.Whish value do you wanna remove?
+echo.---
+echo.Which value do you wanna remove? (might be empty)
 reg query "HKEY_CLASSES_ROOT\Directory\ContextMenus"
+echo...
 echo.
+set value=
 set /p value=
-reg delete "HKEY_CLASSES_ROOT\Directory\ContextMenus\%value%"
+IF NOT "%value%" == "" reg delete "HKEY_CLASSES_ROOT\Directory\ContextMenus\%value%"
+
+echo.Which one of these is meant to be removed with the value before?
 reg query "HKEY_CLASSES_ROOT\Directory\Background\shell"
-echo.Whish one of these is meant to be removed with the value before?
+echo...
 echo.
+set value=
 set /p value=
-reg delete "HKEY_CLASSES_ROOT\Directory\Background\shell\%value%"
+IF NOT "%value%" == "" reg delete "HKEY_CLASSES_ROOT\Directory\Background\shell\%value%"
+
+echo.Which one of these is meant to be removed with the value before?
 reg query "HKEY_CLASSES_ROOT\Directory\shell"
-echo.Whish one of these is meant to be removed with the value before?
+echo...
 echo.
+set value=
 set /p value=
-reg delete "HKEY_CLASSES_ROOT\Directory\shell\%value%"
+IF NOT "%value%" == "" reg delete "HKEY_CLASSES_ROOT\Directory\shell\%value%"
 echo.
 echo.all done!
 pause
 exit
+
+:empty
+echo.you select nothing
+echo.If you dont see what you want, try deleting manually, open "regedit.exe" and
+echo   first go to [HKEY_CLASSES_ROOT\Directory\shell] and remove 01Menu*
+echo   then go to [HKEY_CLASSES_ROOT\Directory\ContextMenus] and remove Menu*
+echo   finally go to [HKEY_CLASSES_ROOT\Directory\backround\shell] and remove Menu*
+echo   and the menu will be removed! 
+pause
